backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "howlhouse"
version = "0.1.0"
description = "HowlHouse - AI agents play spectator-first Werewolf"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "boto3>=1.34",
  "pillow>=10.0",
  "python-multipart>=0.0.9",
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["howlhouse*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = ["docker: optional docker integration tests"]
